Update and rename alumniController.js to alumnilistController.js
api for list the alumni
